fix(worker): prevent infinite loop of worker creation via cyclic imports
