+++ server/integrations/google/utils.ts
@@ -1,4 +1,4 @@
-import type { GaxiosResponse } from "gaxios"
+import type { GaxiosError, GaxiosResponse } from "gaxios"
 import {
   Subsystem,
   type GoogleClient,
@@ -220,6 +220,22 @@ export const getPDFContent = async (
       `Error getting file: ${error} ${(error as Error).stack}`,
       error,
     )
+
+    // previously sync was breaking for these 2 cases
+    // so we return null (TODO: confirm if we ingest atleast the metadata)
+    if (
+      (error as Error).message === "No password given" &&
+      (error as any).code === 1
+    ) {
+      return
+    } else if (
+      (error as Error).message === "Permission denied" &&
+      (error as GaxiosError).code === "EACCES"
+    ) {
+      // this is pdf someone else has shared but we don't have access to download it
+      return
+    }
+
     throw new DownloadDocumentError({
       message: "Error in getting file content",
       cause: error as Error,
@@ -358,111 +374,4 @@ export const checkDownloadsFolder = async (
       `Error checking or deleting files in downloads folder: ${error} ${(error as Error).stack}`,
     )
   }
-}
-
-export async function saveGmailAttachment(
-  attachmentData: any,
-  fileName: string,
-) {
-  try {
-    // The attachment data is base64 encoded, so we need to decode it
-    // Replace any `-` with `+` and `_` with `/` to make it standard base64
-    const normalizedBase64 = attachmentData
-      .replace(/-/g, "+")
-      .replace(/_/g, "/")
-
-    const buffer = Buffer.from(normalizedBase64, "base64")
-    await fs.writeFile(fileName, buffer)
-
-    Logger.info(`Successfully saved gmail attachment at ${fileName}`)
-  } catch (error) {
-    Logger.error("Error saving gmail attachment:", error)
-    throw error
-  }
-}
-
-export const getGmailAttachmentChunks = async (
-  gmail: gmail_v1.Gmail,
-  attachmentMetadata: {
-    messageId: string
-    attachmentId: string
-    filename: string
-    size: number
-  },
-): Promise<string[] | null> => {
-  const { attachmentId, filename, messageId, size } = attachmentMetadata
-  let attachmentChunks: string[] = []
-  const pdfSizeInMB = size / (1024 * 1024)
-  // Ignore the PDF files larger than Max PDF Size
-  if (pdfSizeInMB > MAX_ATTACHMENT_PDF_SIZE) {
-    Logger.warn(
-      `Ignoring ${filename} as its more than ${MAX_ATTACHMENT_PDF_SIZE} MB`,
-    )
-    return null
-  }
-
-  try {
-    const fileName = `${filename}_${messageId}`
-    const downloadAttachmentFilePath = path.join(downloadDir, fileName)
-
-    const attachementResp = await retryWithBackoff(
-      () =>
-        gmail.users.messages.attachments.get({
-          messageId: messageId,
-          id: attachmentId,
-          userId: "me",
-        }),
-      "Fetching Gmail Attachments",
-    )
-
-    await saveGmailAttachment(
-      attachementResp.data.data,
-      downloadAttachmentFilePath,
-    )
-    const docs = await safeLoadPDF(downloadAttachmentFilePath)
-    if (!docs || docs.length === 0) {
-      Logger.warn(`Could not get content for file: ${filename}. Skipping it`)
-
-      await deleteDocument(downloadAttachmentFilePath)
-      return null
-    }
-    attachmentChunks = docs
-      .flatMap((doc) => chunkDocument(doc.pageContent))
-      .map((v) => v.chunk)
-      .filter((v) => v.trim())
-
-    await deleteDocument(downloadAttachmentFilePath)
-  } catch (error) {
-    throw error
-  }
-
-  return attachmentChunks
-}
-
-// Function to parse attachments from the email payload
-export const parseAttachments = (
-  payload: gmail_v1.Schema$MessagePart,
-): { attachments: Attachment[]; filenames: string[] } => {
-  const attachments: Attachment[] = []
-  const filenames: string[] = []
-
-  const traverseParts = (parts: any[]) => {
-    for (const part of parts) {
-      if (part.filename && part.body && part.body.attachmentId) {
-        filenames.push(part.filename)
-        attachments.push({
-          fileType: part.mimeType || "application/octet-stream",
-          fileSize: parseInt(part.body.size, 10) || 0,
-        })
-      } else if (part.parts) {
-        traverseParts(part.parts)
-      }
-    }
-  }
-
-  if (payload.parts) {
-    traverseParts(payload.parts)
-  }
-
-  return { attachments, filenames }
 }
